BackedUpImmutable.jl

Contents:

StaticDict

A Julia Dict that is the ImmutableDict found in Base, but with additional constructors.

BackedUpImmutableDict

A Julia Dict object that has fixed keys, and alterable but restorable values.

Functions Reference

BackedUpImmutable.BackedUpImmutableDict

BackedUpImmutable.BackedUpImmutableDict

BackedUpImmutable.BackedUpImmutableDict

BackedUpImmutable.ImmutableDictError

BackedUpImmutable.StaticDict

Base.ImmutableDict

Base.ImmutableDict

BackedUpImmutable.restore!

BackedUpImmutable.restoreall!

BackedUpImmutable.BackedUpImmutableDict — Type

# BackedUpImmutableDict{K, V} 
* Combines a key, not value, immutable hash dictionary with a backup of the original value defaults.
* For configuration data storage, with a simple restore to default

BackedUpImmutable.BackedUpImmutableDict — Method

Makes a BackedUpImmutableDict from a tuple of key, value pairs (varargs style)

BackedUpImmutable.BackedUpImmutableDict — Method

Makes a BackedUpImmutableDict from a vector of key, value pairs

BackedUpImmutable.ImmutableDictError — Type

ImmutableDictError(msg)

Thrown when an operation would require adding or removing a key from the fixed key set of a StaticDict / BackedUpImmutableDict.

BackedUpImmutable.StaticDict — Type

Another name for ImmutableDict, but here as with an extra constructor.

Base.ImmutableDict — Method

Constructor for StaticDict to take a series of pairs, varargs style

Base.ImmutableDict — Method

Constructor for StaticDict / ImmutableDict to take an array of key value pairs.

BackedUpImmutable.restore! — Method

Restore a key's backed up default value.

BackedUpImmutable.restoreall! — Method

Restore all values back to defaults